\name{celestial-package}
\alias{celestial-package}
\alias{celestial}
\docType{package}
\title{
Celestial coordinate conversion functions.
}
\description{
Various functions for converting between commonly used coordinate systems in astronomy.
}
\details{
\tabular{ll}{
Package: \tab celestial\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2015-01-09\cr
License: \tab GPL-2\cr
}
There are a number of functions included, but the most useful for astronomy conversions are the decimal degrees to DMS/HMS formats used at many telescopes: deg2dms, deg2hms, dms2deg, hms2deg. It also contains functions for various cosmological calculations.
}
\author{
Aaron Robotham

Maintainer: Aaron Robotham <aaron.robotham@uwa.edu.au>
}

\keyword{ package }

