\name{celestial-package}
\alias{celestial-package}
\alias{celestial}
\docType{package}
\title{
Collection of Common Astronomical Conversion Routines and Functions
}
\description{
Various functions for converting between commonly used coordinate systems in astronomy and making cosmological calculations.
}
\details{
\tabular{ll}{
Package: \tab celestial\cr
Type: \tab Package\cr
Version: \tab 1.4.2\cr
Date: \tab 2018-03-27\cr
License: \tab GPL-3\cr
Depends: \tab R (>= 3.00), RANN, NISTunits, pracma\cr
}
There are a number of functions included, but the most useful for astronomy conversions are the decimal degrees to DMS/HMS formats used at many telescopes: deg2dms, deg2hms, dms2deg, hms2deg. It also contains functions for various cosmological calculations (i.e. distance, volume and age for different cosmologies and redshifts).
}
\author{
Aaron Robotham

Maintainer: Aaron Robotham <aaron.robotham@uwa.edu.au>
}

\keyword{ package }
\keyword{ celestial }

